requirements.txt
matplotlib~=3.6.1
PyYAML~=6.0
nordpool~=0.3.3
python-dateutil~=2.8.2
tabulate~=0.9.0
python-dotenv~=0.21.0